prediction_service/requirements.txt
scikit-learn~=1.3.0
pandas
flask
gunicorn
python-dotenv
mlflow
azure-identity
azure-storage-blob
psutil
prometheus_client
pytest
